// src/__generated__/StoryCommentsSectionFragment.graphql.ts
export type StoryCommentsSectionFragment$data = {
  readonly comments: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly " $fragmentSpreads": FragmentRefs<"CommentFragment">;
      } | null | undefined;
    } | null | undefined> | null | undefined;
    readonly pageInfo: {
      readonly endCursor: string | null | undefined;
      readonly hasNextPage: boolean | null | undefined;
      readonly hasPreviousPage: boolean | null | undefined;
      readonly startCursor: string | null | undefined;
    } | null | undefined;
  } | null | undefined;
  readonly id: string;
  readonly " $fragmentSpreads": FragmentRefs<"StoryCommentsComposerFragment">;
  readonly " $fragmentType": "StoryCommentsSectionFragment";
};
export type StoryCommentsSectionFragment$key = {
  readonly " $data"?: StoryCommentsSectionFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"StoryCommentsSectionFragment">;
};